refactor: move primitive array into separate component
